+++ src/main/java/servlets/background/product_type/DeleteServlet.java
@@ -13,6 +13,7 @@ import javax.servlet.http.HttpServletResponse;
 import com.mysql.cj.jdbc.DatabaseMetaData;
 
 import DAO.ProductsClassDAO;
+import entity.Admin;
 
 @WebServlet("/background/pages/function/product_type/delete")
 public class DeleteServlet extends HttpServlet{
@@ -27,14 +28,24 @@ public class DeleteServlet extends HttpServlet{
         req.setCharacterEncoding("UTF-8");
         
         String id = req.getParameter("id");
-        int sqljg = dao.deleteById(id);
-        String jg = "";
-        if (sqljg == -1451) {
-        	jg = "1451";
-		}else if (sqljg>0) {
-        	jg = "true";
-		}
         
-        resp.getWriter().print(jg);
+        Admin admin = (Admin) req.getSession().getAttribute("admin");
+		String adminType = "4";
+	    if (admin != null) {
+			adminType=admin.getAdminTypeId()+"";
+		}
+	    String jg = "";
+	    if (adminType.equals("6")||adminType.equals("1")) {
+	    	 int sqljg = dao.deleteById(id);
+	         if (sqljg == -1451) {
+	         	jg = "1451";
+	 		}else if (sqljg>0) {
+	         	jg = "true";
+	 		}  
+		 }else {
+			jg="权限不足!";
+		}
+
+		resp.getWriter().print(jg);
 	}
 }
